refactor: 게시물 리팩토링
