+++ src/main/kotlin/co/yappuworld/post/infrastructure/PostRepository.kt
@@ -1,6 +1,6 @@
 package co.yappuworld.post.infrastructure
 
-import co.yappuworld.post.domain.PostEntity
+import co.yappuworld.post.infrastructure.entity.PostEntity
 import com.linecorp.kotlinjdsl.support.spring.data.jpa.repository.KotlinJdslJpqlExecutor
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.UUID
