+++ src/main/kotlin/co/yappuworld/user/client/dto/response/AdminSignUpApplicationActivityUnitResponse.kt
@@ -1,7 +1,7 @@
 package co.yappuworld.user.client.dto.response
 
 import co.yappuworld.operation.client.dto.response.PositionResponse
-import co.yappuworld.user.domain.entity.ActivityUnitParam
+import co.yappuworld.user.infrastructure.entity.ActivityUnitParam
 import io.swagger.v3.oas.annotations.media.Schema
 
 data class AdminSignUpApplicationActivityUnitResponse(
