+++ src/main/kotlin/co/yappuworld/user/client/dto/response/LatestSignUpApplicationResponse.kt
@@ -1,6 +1,6 @@
 package co.yappuworld.user.client.dto.response
 
-import co.yappuworld.user.domain.entity.SignUpApplicationEntity
+import co.yappuworld.user.infrastructure.entity.SignUpApplicationEntity
 import co.yappuworld.user.domain.vo.SignUpApplicationStatus
 import io.swagger.v3.oas.annotations.media.Schema
