+++ src/main/kotlin/co/yappuworld/user/client/dto/response/UserAlarmStatusResponse.kt
@@ -1,6 +1,6 @@
 package co.yappuworld.user.client.dto.response
 
-import co.yappuworld.user.domain.entity.UserAlarmSettingEntity
+import co.yappuworld.user.infrastructure.entity.UserAlarmSettingEntity
 import io.swagger.v3.oas.annotations.media.Schema
 
 data class UserAlarmStatusResponse(
