+++ src/main/kotlin/co/yappuworld/user/infrastructure/UserCommandService.kt
@@ -1,9 +1,9 @@
 package co.yappuworld.user.infrastructure
 
-import co.yappuworld.user.domain.entity.SignUpApplicationEntity
-import co.yappuworld.user.domain.entity.UserAlarmSettingEntity
-import co.yappuworld.user.domain.entity.UserDeviceEntity
-import co.yappuworld.user.domain.entity.UserEntity
+import co.yappuworld.user.infrastructure.entity.SignUpApplicationEntity
+import co.yappuworld.user.infrastructure.entity.UserAlarmSettingEntity
+import co.yappuworld.user.infrastructure.entity.UserDeviceEntity
+import co.yappuworld.user.infrastructure.entity.UserEntity
 import co.yappuworld.user.domain.vo.UserRole
 import co.yappuworld.user.infrastructure.jpa.ActivityUnitRepository
 import co.yappuworld.user.infrastructure.jpa.UserAlarmSettingRepository
