+++ src/main/kotlin/co/yappuworld/user/infrastructure/entity/UserDeviceEntity.kt
@@ -1,4 +1,4 @@
-package co.yappuworld.user.domain.entity
+package co.yappuworld.user.infrastructure.entity
 
 import co.yappuworld.global.persistence.BaseEntity
 import jakarta.persistence.Entity
