+++ src/main/kotlin/co/yappuworld/user/infrastructure/jpa/ActivityUnitRepository.kt
@@ -1,6 +1,6 @@
 package co.yappuworld.user.infrastructure.jpa
 
-import co.yappuworld.user.domain.entity.ActivityUnitEntity
+import co.yappuworld.user.infrastructure.entity.ActivityUnitEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.UUID
