+++ src/main/kotlin/co/yappuworld/user/infrastructure/jpa/UserDeviceRepository.kt
@@ -1,6 +1,6 @@
 package co.yappuworld.user.infrastructure.jpa
 
-import co.yappuworld.user.domain.entity.UserDeviceEntity
+import co.yappuworld.user.infrastructure.entity.UserDeviceEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.UUID
